+++ tests/ui/fn_ptr_reference_encode2.rs
@@ -1,7 +1,0 @@
-//! Extra test for `fn_ptr_reference_encode`
-//! (They fail at different compilation passes).
-use objc2::Encode;
-
-fn main() {
-    let encoding = <extern "C" fn(&i32) as Encode>::ENCODING;
-}